// packages/datadog-api-client-v1/models/ProcessQueryDefinition.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* The process query to use in the widget.
*/
export declare class ProcessQueryDefinition {
    /**
    * List of processes.
    */
    'filterBy'?: Array<string>;
    /**
    * Max number of items in the filter list.
    */
    'limit'?: number;
    /**
    * Your chosen metric.
    */
    'metric': string;
    /**
    * Your chosen search term.
    */
    'searchBy'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): ProcessQueryDefinition;
    static serialize(data: ProcessQueryDefinition): {
        [key: string]: any;
    };
    constructor();
}